# Services/ConduitLLM.Gateway/openapi-gateway.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Conduit Gateway API",
    "description": "OpenAI-compatible API for multi-provider LLM access - Requires Bearer token authentication",
    "version": "v1"
  },
  "paths": {
    "/v1/models": {
      "get": {
        "tags": [
          "Models"
        ],
        "operationId": "Models_ListModels",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAIErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "VirtualKey": [ ]
          }
        ]
      }
    },
    "/v1/models/{modelId}/metadata": {
      "get": {
        "tags": [
          "Models"
        ],
        "operationId": "Models_GetModelMetadata",
        "parameters": [
          {
            "name": "modelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelMetadataResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAIErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAIErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "VirtualKey": [ ]
          }
        ]
      }
    },
...
    "/v1/discovery/models": {
      "get": {
        "tags": [
          "Discovery"
        ],
        "operationId": "Discovery_GetModels",
        "parameters": [
          {
            "name": "capability",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoveryModelsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "VirtualKey": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DiscoveredModelDto": {
        "required": [
          "id",
          "provider",
          "display_name",
          "description",
          "model_card_url",
          "max_tokens",
          "max_input_tokens",
          "max_output_tokens",
          "tokenizer_type",
          "input_modalities",
          "output_modalities",
          "capability_source",
          "capabilities_last_verified_at",
          "parameters",
          "capabilities"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "provider": {
            "type": [
              "null",
              "string"
            ]
          },
          "display_name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "model_card_url": {
            "type": "string"
          },
          "max_tokens": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "max_input_tokens": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "max_output_tokens": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "tokenizer_type": {
            "type": "string"
          },
          "input_modalities": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "output_modalities": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "capability_source": {
            "type": "string"
          },
          "capabilities_last_verified_at": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "parameters": {
            "type": "string"
          },
          "capabilities": {
            "$ref": "#/components/schemas/ModelCapabilitiesDto"
          }
        },
        "description": "A model returned by the Conduit discovery extension."
      },
...
      "DiscoveryModelsResponse": {
        "required": [
          "data",
          "count"
        ],
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DiscoveredModelDto"
            }
          },
          "count": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        },
        "description": "Model discovery response."
      },
...
      "ModelCapabilitiesDto": {
        "required": [
          "chat",
          "chat_stream",
          "image_input",
          "video_input",
          "audio_input",
          "file_input",
          "vision",
          "video_understanding",
          "image_generation",
          "video_generation",
          "embeddings",
          "function_calling",
          "speech_to_text",
          "text_to_speech",
          "rerank"
        ],
        "type": "object",
        "properties": {
          "chat": {
            "type": "boolean"
          },
          "chat_stream": {
            "type": "boolean"
          },
          "image_input": {
            "type": "boolean"
          },
          "video_input": {
            "type": "boolean"
          },
          "audio_input": {
            "type": "boolean"
          },
          "file_input": {
            "type": "boolean"
          },
          "vision": {
            "type": "boolean"
          },
          "video_understanding": {
            "type": "boolean"
          },
          "image_generation": {
            "type": "boolean"
          },
          "video_generation": {
            "type": "boolean"
          },
          "embeddings": {
            "type": "boolean"
          },
          "function_calling": {
            "type": "boolean"
          },
          "speech_to_text": {
            "type": "boolean"
          },
          "text_to_speech": {
            "type": "boolean"
          },
          "rerank": {
            "type": "boolean"
          },
          "tool_use": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "json_mode": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "max_tokens": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "max_output_tokens": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          }
        },
        "description": "Capabilities advertised for a model."
      },
      "ModelListItemDto": {
        "required": [
          "id",
          "object"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "object": {
            "type": "string"
          }
        },
        "description": "An OpenAI-compatible model entry."
      },
      "ModelListResponse": {
        "required": [
          "data",
          "object"
        ],
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModelListItemDto"
            }
          },
          "object": {
            "type": "string"
          }
        },
        "description": "An OpenAI-compatible model list."
      },
      "ModelMetadataDto": {
        "required": [
          "id",
          "canonical_model_id",
          "canonical_name",
          "provider",
          "provider_model_id",
          "description",
          "model_card_url",
          "input_modalities",
          "output_modalities",
          "capability_source",
          "capabilities_last_verified_at",
          "capabilities",
          "max_input_tokens",
          "max_output_tokens"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "canonical_model_id": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "canonical_name": {
            "type": "string"
          },
          "provider": {
            "type": [
              "null",
              "string"
            ]
          },
          "provider_model_id": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "model_card_url": {
            "type": [
              "null",
              "string"
            ]
          },
          "input_modalities": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "output_modalities": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "capability_source": {
            "type": "string"
          },
          "capabilities_last_verified_at": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "capabilities": {
            "$ref": "#/components/schemas/ModelCapabilitiesDto"
          },
          "max_input_tokens": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "max_output_tokens": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          }
        },
        "description": "Effective metadata for an enabled model mapping."
      },
      "ModelMetadataResponse": {
        "required": [
          "modelId",
          "metadata"
        ],
        "type": "object",
        "properties": {
          "modelId": {
            "type": "string"
          },
          "metadata": {
            "$ref": "#/components/schemas/ModelMetadataDto"
          }
        },
        "description": "Model metadata response envelope."
      },
...
      "OpenAIError": {
        "required": [
          "message",
          "type"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "param": {
            "type": [
              "null",
              "string"
            ]
          },
          "code": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "OpenAIErrorResponse": {
        "required": [
          "error"
        ],
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/OpenAIError"
          }
        }
      },
...
    },
    "securitySchemes": {
      "VirtualKey": {
        "type": "http",
        "description": "ConduitLLM virtual key supplied as an opaque Bearer token.",
        "scheme": "bearer",
        "bearerFormat": "opaque virtual key"
      }
    }
  },
...
}